refactor(vite): split server-only into ast-walkers, source-extraction, stub-templates (#24)
